+++ flask_typed_routes/app.py
@@ -53,6 +53,8 @@ class FlaskTypedRoutes:
         by default it uses the default error handler provided by the library.
     :param Tuple[str] ignore_verbs: HTTP verbs to ignore.
     :param str mode: Mode of operation, 'auto' or 'manual'. Default is 'auto'.
+    :param str exclude_doc_url_prefix: Exclude the OpenAPI documentation URL prefix.
+    :param dict openapi: OpenAPI schema definition for the application.
     """
 
     IGNORE_VERBS = ("HEAD", "OPTIONS")
@@ -63,6 +65,7 @@ class FlaskTypedRoutes:
         validation_error_handler=None,
         ignore_verbs=None,
         mode=Mode.auto,
+        exclude_doc_url_prefix=None,
         **openapi,
     ):
         self.error_handler = validation_error_handler
@@ -70,6 +73,7 @@ class FlaskTypedRoutes:
         if mode not in (Mode.auto, Mode.manual):
             raise ValueError(f"Invalid mode: {mode}")
         self.mode = mode
+        self.api_doc_prefix = exclude_doc_url_prefix
         self.openapi_schema = ftr_openapi.get_openapi(**openapi)
         if app:
             self.init_app(app)
@@ -91,31 +95,34 @@ class FlaskTypedRoutes:
         @functools.wraps(func)
         def wrapper(rule, endpoint=None, view_func=None, **kwargs):
             path_args = ftr_utils.extract_rule_params(rule)
-            if view_func:
-                view = ftr_utils.class_based_view(view_func)
-
-                if view:  # class-based view
-                    verbs = view.methods or ()  # methods defined in the class
-                    verbs = (verb for verb in verbs if hasattr(view, verb.lower()))  # implemented methods
+            is_apidoc = self.api_doc_prefix and rule.startswith(self.api_doc_prefix)
+            if view_func and not is_apidoc:
+                # name of the view function or view class if not endpoint is provided
+                view_name = endpoint or view_func.__name__
+                view_class = ftr_utils.class_based_view(view_func)
+                if view_class:  # class-based view
+                    verbs = view_class.methods or ()  # methods defined in the class
+                    verbs = (verb for verb in verbs if hasattr(view_class, verb.lower()))  # implemented methods
                     verbs = frozenset(verbs).difference(self.ignore_verbs)  # ignore some methods
 
                     if verbs:  # implemented methods
                         for verb in verbs:
-                            method = getattr(view, verb.lower())
+                            method = getattr(view_class, verb.lower())
                             if self.is_typed(method):
                                 new_method = ftr_core.route(method, path_args)
-                                self.merge_openapi_route(new_method, rule, endpoint, kwargs)
-                                setattr(view, verb.lower(), new_method)
+                                new_kwargs = {**kwargs, "methods": [verb]}
+                                self.register_openapi_route(new_method, rule, view_name, new_kwargs, path_args)
+                                setattr(view_class, verb.lower(), new_method)
 
                     # no implemented methods, use the default "dispatch_request"
-                    elif self.is_typed(view.dispatch_request):
-                        new_method = ftr_core.route(view.dispatch_request, path_args)
-                        self.merge_openapi_route(new_method, rule, endpoint, kwargs)
-                        view.dispatch_request = new_method
+                    elif self.is_typed(view_class.dispatch_request):
+                        new_method = ftr_core.route(view_class.dispatch_request, path_args)
+                        self.register_openapi_route(new_method, rule, view_name, kwargs, path_args)
+                        view_class.dispatch_request = new_method
 
                 elif self.is_typed(view_func):  # function-based view
                     view_func = ftr_core.route(view_func, path_args)
-                    self.merge_openapi_route(view_func, rule, endpoint, kwargs)
+                    self.register_openapi_route(view_func, rule, view_name, kwargs, path_args)
 
             return func(rule, endpoint=endpoint, view_func=view_func, **kwargs)
 
@@ -125,19 +132,18 @@ class FlaskTypedRoutes:
         enabled = getattr(view_func, ftr_utils.TYPED_ROUTE_ENABLED, False)
         return self.mode == Mode.auto or enabled
 
-    def merge_openapi_route(self, func, rule, endpoint, kwargs, /):
-        methods = kwargs.get("methods") or getattr(func, "methods", ()) or ("GET",)
-        endpoint = endpoint or func.__name__
-        spec = ftr_openapi.get_operations(func, rule, endpoint, methods)
-        self.merge_openapi_paths(spec["paths"])
-        self.merge_openapi_schemas(spec["components"]["schemas"])
+    def register_openapi_route(self, func, rule, func_name, kwargs, path_args, /):
+        methods = kwargs.get("methods") or ("GET",)
+        spec = ftr_openapi.get_operations(func, rule, func_name, methods, path_args)
+        self.register_openapi_paths(spec["paths"])
+        self.register_openapi_schemas(spec["components"]["schemas"])
 
-    def merge_openapi_paths(self, paths, /):
+    def register_openapi_paths(self, paths, /):
         current = self.openapi_schema["paths"]
         for path, spec in paths.items():
             current[path].update(spec)
 
-    def merge_openapi_schemas(self, schemas, /):
+    def register_openapi_schemas(self, schemas, /):
         merged = self.openapi_schema["components"]["schemas"]
         for name, schema in schemas.items():
             matches = (mg_name for mg_name, mg_schema in merged.items() if mg_schema == schema)